Can't use security providers
The Oracle JVM, the primary one on windows, is very finicky
about SecurityProviders, and requires modifications to allow
more than 128bit encryption. We can't expect people to change
their JVM, we're just going to suck it up. 256bit is still
an option for messages, we just won't generate them here